+++ extensions/brave-search-with-results/src/hooks/useHistory.ts
@@ -1,11 +1,12 @@
 import { LocalStorage } from "@raycast/api";
 import { randomUUID } from "crypto";
-import { useCallback, useEffect, useState } from "react";
+import { useCallback, useEffect, useMemo, useState } from "react";
 
 const STORAGE_KEY = "history";
 
 export enum Type {
   Web,
+  Image,
 }
 
 export interface HistoryItem {
@@ -18,12 +19,12 @@ export interface HistoryItem {
 const get = async () => JSON.parse((await LocalStorage.getItem(STORAGE_KEY)) ?? JSON.stringify([])) as HistoryItem[];
 const set = (history: HistoryItem[]) => LocalStorage.setItem(STORAGE_KEY, JSON.stringify(history));
 
-export default function useHistory() {
+export default function useHistory(type: Type) {
   const [isLoading, setIsLoading] = useState(false);
   const [items, setItems] = useState([] as HistoryItem[]);
 
   const addToHistory = useCallback(
-    (query: string, type: Type) => {
+    (query: string) => {
       const newItem = {
         id: randomUUID(),
         query: query.toLocaleLowerCase(),
@@ -39,7 +40,7 @@ export default function useHistory() {
       setItems(nextItems);
       return set(nextItems);
     },
-    [items, setItems],
+    [items, setItems, type],
   );
 
   const removeFromHistory = useCallback(
@@ -69,5 +70,9 @@ export default function useHistory() {
     initialize();
   }, []);
 
-  return { isLoading, items, addToHistory, removeFromHistory, clearHistory };
+  const filteredItems = useMemo(() => {
+    return items.filter((item) => item.type === type);
+  }, [items, type]);
+
+  return { isLoading, items: filteredItems, addToHistory, removeFromHistory, clearHistory };
 }