+++ extensions/brave-search-with-results/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Brave Search with Results Changelog
 
+## 2025-07-21
+
+- Image search
+
 ## [Initial Version] - 2024-05-04
 
 - Web results